requirements.txt
jax==0.4.13
jaxlib==0.4.13
optax==0.1.7
ipympl==0.9.3
matplotlib==3.7.1
hydra-core==1.3.2
hydra-joblib-launcher==1.2.0
numpy==1.24
pandas==2.0.1
seaborn==0.12.1
scipy==1.10.1
tqdm==4.65.0
numba==0.57.1
bayesnewton==1.3.4
objax==1.7.0
jaxopt==0.8
tikzplotlib
jupyterlab